feat(workflow): remove 'collected' field from Job and HistoricalJob models feat(workflow): add 'coolected' field to Job and HistoricalJob models feat(workflow): remove 'password_needs_reset' field from Staff model feat(workflow): add 'password_needs_reset' field to Staff model
The 'collected' field was removed from the Job and HistoricalJob models,
and a new 'coolected' field was added. This was likely done to correct
a typo in the field name. The 'password_needs_reset' field was removed
from the Staff model and then added back. This was likely done to
resolve an issue with makemigrations not working.