feat(add_client.html): display error messages in a styled alert box
This commit introduces a new styled alert box to display error
messages when adding a new client. The alert box includes a heading,
a general error message, and optional details such as the error type,
the name of the client that failed to be added, and the client's
email address. The alert box is initially hidden and is displayed
when an error occurs. This provides a more user-friendly way to
display error messages to the user.