feat(settings): configure django logging to output to console and file
This commit configures the django logging settings to output logs to both
the console and a file. This change was made to improve the visibility
of logs and make it easier to debug issues.
The log level for the console handler was changed to DEBUG to provide
more detailed information. A verbose formatter was added to the console
handler to include more information in the log messages.
The django.db.backends logger was configured to use both the sql_file
and console handlers.
The xero logger was configured to use both the xero_file and console
handlers.