feat(job): update purchase order line reference in MaterialEntry
This commit updates the foreign key reference in the MaterialEntry
model to point to 'purchasing.purchaseorderline'. This ensures
that the job app correctly references the purchase order line
model in the purchasing app after the purchasing models were moved
to a separate database. A SeparateDatabaseAndState migration is
used to perform a state operation without any database operations.