feat(purchasing): move purchasing models to purchasing app
This commit moves the following models from the workflow app to the
purchasing app:

- PurchaseOrder
- PurchaseOrderLine
- PurchaseOrderSupplierQuote
- Stock

This change is done to improve the organization of the codebase and
to better reflect the domain of the models. The purchasing app is
the natural home for these models, as they are all related to
purchasing activities.

This commit also creates proxy models in the workflow app for the
moved models. This is done to maintain backward compatibility for
existing code that imports from workflow.