refactor(purchasing): rename 'edit_purchase_order' url to 'purchase_orders_detail'
The URL name 'edit_purchase_order' was misleading as the view
associated with it displays the details of a purchase order, rather
than providing a form to edit it. This commit renames the URL to
'purchase_orders_detail' to more accurately reflect its purpose.
This change involves updating the URL pattern in `purchasing/urls.py`
and updating references to the URL in `apps/job/views/edit_job_view_ajax.py`
and `apps/purchasing/views/purchase_order.py`.