Please read how the KPI report calculates for a day.
I'd like to make a tweak to how quotes are calculated.

We talk about time, materials, and adjustments.  I'd like to make a fourth entry: quote adjustment.

On the day a quoted job is invoiced, I'd like to create that entry with the value being the difference between the quoted price of the job and the reality price of the job.  Note the quoted cost is irrelevant, this just affects revenue.